requirements.txt
typer==0.12.5
rich==13.9.4
faiss-cpu==1.12.0
numpy>=2.1,<3
tiktoken>=0.8.0
fastembed>=0.4.0
python-dotenv==1.0.1
openai==1.107.0
